all user account transaction

+++ Bus-Ticketing-System-API/API/Controllers/adminCuponController.cs
@@ -49,6 +49,10 @@ namespace API.Controllers
         {
             try
             {
+                if (adminCuponService.GetDiscountCupon(obj.id) == null)
+                {
+                    return Request.CreateResponse(HttpStatusCode.Forbidden, new { message = "The cupon doesn't exits" });
+                }
                 obj.admin_id = getID(Request);
                 var data = adminCuponService.updateDiscountCupon(obj);
                 string message = data ? "Cupon is updated" : "Cupon is not updated";
@@ -66,6 +70,10 @@ namespace API.Controllers
         {
             try
             {
+                if(adminCuponService.GetDiscountCupon(id) == null)
+                {
+                    return Request.CreateResponse(HttpStatusCode.Forbidden, new { message = "The cupon doesn't exits" });
+                }
                 var data = adminCuponService.deleteDiscountCupon(id);
                 string message = data ? "Cupon is deleted" : "Cupon is not deleted";
                 return Request.CreateResponse(HttpStatusCode.OK, new { message = message});
@@ -82,6 +90,10 @@ namespace API.Controllers
         {
             try
             {
+                if (adminCuponService.GetDiscountCupon(id) == null)
+                {
+                    return Request.CreateResponse(HttpStatusCode.Forbidden, new { message = "The cupon doesn't exits" });
+                }
                 var data = adminCuponService.GetDiscountCupon(id);
                 //string message = data ? "Cupon is deleted" : "Cupon is not deleted";
                 return Request.CreateResponse(HttpStatusCode.OK, data);
