+++ Bus-Ticketing-System-API/API/Controllers/employeeNoticeController.cs
@@ -49,6 +49,10 @@ namespace API.Controllers
         {
             try
             {
+                if (employeeNoticeService.GetNotice(obj.id) == null)
+                {
+                    return Request.CreateResponse(HttpStatusCode.Forbidden, new { message = "The bus provider doesn't exits" });
+                }
                 obj.emp_id = getID(Request);
                 var data = employeeNoticeService.updateNotice(obj);
                 string message = data ? "notice is updated" : "notice is not updated";
@@ -66,6 +70,10 @@ namespace API.Controllers
         {
             try
             {
+                if (employeeNoticeService.GetNotice(id) == null)
+                {
+                    return Request.CreateResponse(HttpStatusCode.Forbidden, new { message = "The bus provider doesn't exits" });
+                }
                 var data = employeeNoticeService.deleteNotice(id);
                 string message = data ? "notice is deleted" : "notice is not deleted";
                 return Request.CreateResponse(HttpStatusCode.OK, new { message = message });
@@ -82,6 +90,10 @@ namespace API.Controllers
         {
             try
             {
+                if (employeeNoticeService.GetNotice(id) == null)
+                {
+                    return Request.CreateResponse(HttpStatusCode.Forbidden, new { message = "The bus provider doesn't exits" });
+                }
                 var data = employeeNoticeService.GetNotice(id);
                 //string message = data ? "notice is deleted" : "notice is not deleted";
                 return Request.CreateResponse(HttpStatusCode.OK, data);
